requirements.txt
fastapi
uvicorn
opencv-python
numpy
torch
torchvision
Pillow
python-multipart
